feat: split cose and jwt into independent default-on features
Make cose-rust, openssl, and jsonwebtoken optional behind two
independent default-on features:

  cose = [dep:cose-rust, dep:openssl, dep:jsonwebtoken]
  jwt  = [dep:jsonwebtoken]

cose no longer implies the jwt feature -- it depends on the
jsonwebtoken *crate* (used by from_cose_jwk to parse JWK keys), not the
jwt feature's API surface. This lets JWT-only builds drop cose+openssl
and COSE-only builds drop the jwt feature's code, with the two features
genuinely orthogonal.

The jsonwebtoken import in ear.rs is now gated on any(cose, jwt) so the
shared JWK types are available to the cose path without forcing the jwt
feature on. JWT-only doctests are wrapped in cfg(feature = "jwt")
blocks so the feature matrix (cose-only / jwt-only / default) all pass
cargo test cleanly.
