+++ Cargo.toml
@@ -9,15 +9,23 @@ license = "Apache-2.0"
 keywords = ["ear", "serde"]
 categories = ["data-structures", "encoding"]
 
+
 [dependencies]
 base64 = "0.22.1"
 ciborium = "0.2.0"
-cose-rust = "0.1.2"
+cose-rust = {version = "0.1.2", optional = true}
 hex = "0.4.3"
-jsonwebtoken = { version = "10", features = ["aws_lc_rs"] }
+jsonwebtoken = {version = "10", features = ["aws_lc_rs"], optional = true}
 lazy_static = "1.5.0"
-openssl = "0.10.54"
+openssl = {version = "0.10.54", optional = true}
 phf = {version = "0.13.1", features = ["macros", "serde"]}
 serde = {version = "1.0", features = ["derive"]}
 serde_json = {version = "1.0.93", features = ["raw_value"]}
 thiserror = "2.0.16"
+
+
+[features]
+default = ["cose", "jwt"]
+cose = ["dep:cose-rust", "dep:openssl", "dep:jsonwebtoken"]
+jwt = ["dep:jsonwebtoken"]
+
